requirements.txt
cryptography==2.7
numpy==1.16.2
azure-storage==0.36.0